backend/requirements.txt
torch
transformers
Flask==1.0
python-docx
PyPDF2